api/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlmodel>=0.0.22,<1.0
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
python-multipart>=0.0.9,<1.0
pypdf>=5.0,<6.0
python-docx>=1.1,<2.0
rapidfuzz>=3.9,<4.0
httpx>=0.27,<1.0
orjson>=3.10,<4.0
python-dateutil>=2.9,<3.0
pytest>=8.0,<9.0
